let data-parsley attributes have preference over html5 attributes
the reason is that html5 attributes might be necessary for whatever
reason, but data-parsley-* attributes are only used for parsley
validation.

Example: a mobile oriented site might want to set an input's type to "number", so that the UA shows a numeric keyboard, but at the same "number" validation might not be desired.